requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt[crypto]==5.3.1
django-cors-headers==4.6.0
django-extensions==3.2.3
django-filter==24.3
django-silk==5.3.2
django-redis==5.4.0
drf-spectacular==0.28.0
psycopg[c]==3.2.3
pillow==11.0.0
tzdata==2024.2
celery==5.4.0
redis==5.2.1
# redis[hiredis]==5.2.1
uwsgi==2.0.28